Fix wata payment status resolver for miniapp
